Quiet logs for normal operation
The standard behavior is to log the full details of every job. In the
case of some events, this is a rather large JSON payload, making the
logs very unwieldy. This commit changes the log level from NOTICE to
INFO and adds a toString signature to get a summary of the job without
the full payload. The full details are logged at the DEBUG level.